READMEと基本例のコメントを修正し、クエリビルダーを追加。TFIDFVectorizerのデフォルト型をf16に変更し、シリアライズ機能を強化。類似度計算メソッドを更新し、クエリ処理を最適化。

+++ src/vectorizer/mod.rs
@@ -7,6 +7,7 @@ pub mod evaluate;
 use std::{rc::Rc, sync::Arc};
 use std::hash::Hash;
 
+use half::f16;
 use num_traits::Num;
 use ::serde::{Deserialize, Serialize};
 
@@ -16,7 +17,7 @@ use crate::{utils::{datastruct::{vector::{ZeroSpVec, ZeroSpVecTrait}}, normalize
 pub type KeyRc<K> = Rc<K>;
 
 #[derive(Debug, Clone)]
-pub struct TFIDFVectorizer<N = f32, K = String, E = DefaultTFIDFEngine>
+pub struct TFIDFVectorizer<N = f16, K = String, E = DefaultTFIDFEngine>
 where
     N: Num + Copy + Into<f64> + Send + Sync,
     E: TFIDFEngine<N, K> + Send + Sync,
